Provide AM/FM radio QML plugin 0.1
qml-radio-plugin is a standalone QML plugin building on top
of Qt5/QML and RTL-SDR (RTL2832U chipsets ; it is planned
to also support different hardware). It provides a simple
API able to detect, initialize and playback radio from QML
applications. It also ships with an example.

(Qt itself already provides a Radio class in QtMultimedia,
but it has little to no implementations, and is far too
much demanding for a start. Let us do this first)

Change-Id: Ic6d2417247f7227ccb84d4e1d9f8b1c58319390f
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>