Add a QML plugin to expose AMB's ClimateControl objects.
To expose the AMB ClimateControl objects to the application layer, add a
QML plugin so the application developers can natively configure the
vehicle's climate controls based on user input.

Because the potential compatibility issue between AMB and Qt being
discussed on the AGL mailing list has not yet been resolved, this commit
also adds a helper program that the climate control plugin will use as a
workaround to access AMB that may be removed in the future.

To include support for this plugin, it is necessary for users to add
"climatecontrolplugin" to their conf/local.conf when building the image.

Change-Id: I6117f0a13e4195e460e3b552befb6e326cdf0f6a
Signed-off-by: Matthew Vick <mvick@jaguarlandrover.com>