Fix qml-radio-plugin compilation with security flags
Since we added compilation security flags (https://gerrit.
automotivelinux.org/gerrit/#/c/5631/), qml-radio-plugin
breaks because it relies on libtool, which assumes we
want to create an executable due to the flags (-pie)
being passed.

The bug has been reported upstream:
https://lists.gnu.org/archive/html/bug-libtool/2016-05/
msg00004.html
but let us work around it since it is a demo target, and
an upstream fix will take time to implement and test.

Change-Id: I7018df928eb23bc2e8b6afb1aafea46958fd523e
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>