openivi-html5.sh: Configure HTML5 application
Select HTML5 application that should be loaded on
boot (such as a homescreen) through the systemd
service of openivi-html5. Following the
recommendations of XDG base dir specification,
the configuration file should be located at
$XDG_CONFIG_DIRS/openivi-html5/openivi-html5.ini
and it should contain value for homescreen with
path to HTML file. If the configuration file does
not exist or if the HTML does not exist then
openivi-html5 will load the following sample
HTML5 by default:
/usr/share/openivi/example/cluster/index.html

Bug-AGL: SPEC-332

Change-Id: I496b0ccbcb09f8964042f9d07abad2d552376b80
Signed-off-by: Leon Anavi <leon.anavi@konsulko.com>