Add recipe of AGL Demo HMI (Media Player)
This recipe is to deliver AGL Demo Platform for
CES2017 AGL Showcase by bitbake integration.

To use hmi of Media Player, it should be installed on the
target board at least once afer booting.

The package of application, mediaplayer.wgt would
be installed in /usr/AGL/apps on the target board.

To install it, follow these steps from the target console
or remote shell.
  # cd /usr/AGL/apps
  # afm-util install mediaplayer.wgt

Change-Id: I30e69e0027c10a654f657fea250b8f76033f5a52
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>