Add recipe of AGL Demo HMI (HVAC)
This recipe is to deliver AGL Demo Platform for
CES2017 AGL Showcase by bitbake integration.

To use hmi of HVAC,  it should be installed on the target
board at least once afer booting.

The package of application, hvac.wgt would be installed
in /usr/AGL/apps on the target board.

To install it, follow these steps from the target console
or remote shell.
  # cd /usr/AGL/apps
  # afm-util install hvac.wgt

Change-Id: I4d90dcae44fc6a5bedf2b26682ae2dcbfd4db6ee
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>