Add recipe of AGL Demo HMI (Settings)
This recipe is to deliver AGL Demo Platform for
CES2017 AGL Showcase by bitbake integration.

To use hmi of Settings(Date/Wifi/Bluetooth), it should be
installed on the target board at least once afer booting.

The package of application, settings.wgt would be
installed in /usr/AGL/apps on the target board.

To install it, follow these steps from the target console
or remote shell.
      # cd /usr/AGL/apps
      # afm-util install settings.wgt

Change-Id: Ice204d7fd002f4c25f34ae70b38607dc23d85f44
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>