Add recipe for mapviewer
This patch addes recipe for mapviewer.
mapviewer is a child application which communicates
with navigation app on homescreen. In CES2017 demo
this app is launched at virtual screen next to homescreen.
(invisible on homescreen) Mapviewer dislay on virtual screen is
transported to meter cluster side via ethernet.
That will be shown at center of meter cluster with 384x368.

window manager tries to change the position to be sent to a cluster
after systemd execute LayerManagerControl set surface.
In order to avoid this sleep is added as a tentative solution.

Change-Id: I903ef19b2550b01e1c90e8960d1edda01b46297d

Signed-off-by: ynakamura <ynakamura@jp.adit-jv.com>
Signed-off-by: Naoto Yamaguchi <i33399_YAMAGUCHI@aisin-aw.co.jp>