Fix issue of media player which doesn't play next while background
Current window manager make own ivi-layer to each application
and adjust the order when rendering (screen's render order of
layers).

E.g. When showing apps launcher, window manager has only 1
ivi-layer for homescreen under screen 0. When showing an app,
window manager add 1 ivi-layer for it and set render order
of 2 layers, front: app's and back: homescreen layer.
And when app hides, it's layer is removed from render order.

But if removed from render order, excecuting QML also stopped.
It is caused why the media player doesn't play next song
when it is not shown.

This patch changes how to make render order to keep media
player's layer in render order even when it is not shown.

This commit is modification of recipe for homescreen which
contains 1 patch file to 'staging/HomeScreen' because
it is already locked (It is splitted new repositories
but chinook doesn't use it).

Bug-AGL: SPEC-395

Change-Id: I4b1ed16c5d20ccab9a1a2aa55e463d0141d3a170
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>