Remove unused controls app from default install
The "controls" demo app is not particularly useful at this point, and
where as it was not launchable from the homescreen in Chinook, the new
dynamic list in Dab picks it up when it is installed, so it both throws
off the current app icon spacing and makes thing look odd because it
has no icon.  So, just remove it to clean up the homescreen.

Change-Id: I617e832b559b4df54e197f0056cb9da84fed4d7b
Signed-off-by: Scott Murray <scott.murray@konsulko.com>