fix MOST driver dependency issues
This patch adds a new "most" receipe for building the MOST driver. The
receipe replaces the old ones, as it builds all modules at once. Hence,
the receipes aim*, hdm* and mostcore are removed. Additionally, the *.bb
file for the demo platform is modified to reflect these changes.

This patch is needed to avoid dependency issues among the modules of the
driver stack. Although, the problem has been reported with SPEC 438 and a
patch set to fix it has been introduced with Change #8447, it is still
present on RaspberryPi.

Change-Id: Id6104ee6d9bf1b7d9744761a6529461176fdd8b2
Signed-off-by: Christian Gromm <christian.gromm@microchip.com>