packagegroup-agl-appfw-graphics: Add opencv
We add support for a relatively minimal opencv library to the
appfw-grahpics packagegroup.  As part of this, we set the PACKAGECONFIG
for opencv to a minimal level that enables support for things we already
have in the system as well as using V4L devices.  Additional features
that are normally enabled in OpenEmbedded are left out.

Bug-AGL: SPEC-912

Change-Id: Iae4480420ccb619ef79f45de30c86e05831e70f9
Signed-off-by: zheng_wenlong <wenlong_zheng@nexty-ele.com>
Signed-off-by: Tom Rini <trini@konsulko.com>