2nd part of the layer/profile rework [2/2]
Main goal of this is to move the agl-service-* bindings into the core and to
unify common recipes like wayland/weston into meta-agl-profile-graphical.
As a placeholder for recipes is meta-agl-profile-demo which are being reshuffled.

TLDR: remove recipes from meta-agl-demo that were moved to other layers

Bug-AGL: SPEC-145 , SPEC-876 , SPEC-877

v2: moving updated udisks changes to meta-agl

Change-Id: I873107f00adadeb6f3f266cce37b513fe6d44cf2
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>