Add agl-demo-preload feature
Add a feature to preload the maps for the navigation app and the API key
for the POI app.  The intent is to allow building an image that works
"out of the box" without manually installing things.  The selection of
UK or Japan maps can be controlled by setting the variable
DEMO_MAPS_LOCALE to either "uk" (default) or "jp" in local.conf or
site.conf.  The POI API key should be set in one of those locations with
the variable POIAPP_API_KEY (default value is an empty string).

Change-Id: I25537935ddac740090f2121d0a83305c746185b0
Signed-off-by: Scott Murray <scott.murray@konsulko.com>