poiapp-api-key: Add client ID
Somehow in my original testing the POI app was working even though
the client ID was not being added to /etc/poikey as the first line.
Add a POIAPP_CLIENT_ID variable to go with POIAPP_API_KEY, and write
them both to the file. As well, mark the package as platform
independent, as it contains just a text configuration file.

Change-Id: Iab5d6aa44007c61457697882d9a2ece12f984acf
Signed-off-by: Scott Murray <scott.murray@konsulko.com>