Disable btwilink by default on m3ulcb
Testing for CES has found the wilink BT on the M3+Kingfisher somewhat
unreliable.  To disable it by default for now, add a demo platform
configuration hook to packagegroup-agl-demo-platform and use it to
pick up a modprobe configuration file to prevent btwilink from loading.
This addition is specific to the m3ulcb platform, and the variable
DEMO_ENABLE_BTWILINK can be used to over-ride the behavior by setting
it to "true" in local.conf or site.conf.

Change-Id: Ia645d9329594547f83601526e3f3fd5da5845b24
Signed-off-by: Scott Murray <scott.murray@konsulko.com>