navigation: Add 4A support
To properly support 4A in the navigation app, add a patch to pull in
the 4A binding, add some code to open / close the "navigation" role,
and use a simple gstreamer pipeline to play the generated files.  The
existing playback script templates have had their playback commands
removed, but are retained for driving the voice file generation.

Note that a patch is being used for now, as gpsnavi is managed outside
of AGL on github.  Additionally, that repository has no flounder branch,
and its HEAD does not work on flounder at present, complicating things.

Change-Id: Ie5c63c2ae22237a80dc034b98d7279d3c203273f
Signed-off-by: Scott Murray <scott.murray@konsulko.com>