agl-demo-platform-crosssdk: Add mosquitto-dev
Add mosquitto-dev to agl-demo-platform-crosssdk to enable CI to
build the telematics demo image application.  This is likely a
stopgap for now, as the plan to handle this situation going forward
(i.e. use separate profile SDKs versus one superset one) will be
discussed at the May 2019 F2F.  A comment has been added to indicate
this.

Change-Id: Ibd48842d95c246dc0b7e7f0ec3fdb8300da55f7b
Signed-off-by: Scott Murray <scott.murray@konsulko.com>