Remove qtaglextras and dependencies
Remove qtaglextras and its dependencies libhomescreen and
agl-service-homescreen, as they have been moved to the
meta-agl-profile-graphical-qt5 and meta-agl-profile-graphical
layers in meta-agl.  This is required to fix building the
cluster demo SDK now that qtaglextras has been added to
packagegroup-qt5-toolchain-target.  These changes are a
partial backport of the move of the contents of
meta-hmi-framework in master / Halibut.

Bug-AGL: SPEC-2484

Change-Id: I057c863a4fd2c87d419c18b114a6c4473c4d75c3
Signed-off-by: Scott Murray <scott.murray@konsulko.com>