Update demo features for switch to ondemandnavi
The ondemandnavi navigation application does not work with the
mapviewer cluster demo application, so update the logic to handle
the agl-cluster-demo-support feature in packagegroup-agl-demo-platform
to not pull it in when using ondemandnavi as the virtual/navigation
provider.  Since the plan is to replace mapviewer with an equivalent
QtLocation based map streamer for the cluster demo, the mapviewer-demo
recipe has also been replaced with a more generically named
cluster-demo-network-config recipe.  Additionally, the agl-demo-preload
feature logic has been tweaked to not pull in the map package for the
old navigation application when using ondemandnavi.

Bug-AGL: SPEC-2576

Change-Id: I57b45ab152f9fab6cee97d570be1949136cd39eb
Signed-off-by: Scott Murray <scott.murray@konsulko.com>