radio: Update default presets location
Since the changes to run the applications as non-root, the location
of the default preset file has not matched with the agl-driver user
the application runs as.  Since simply changing the install location
to /home/1001 would require some post-install SMACK label tinkering,
and is perhaps fragile to the application user, the files are now
instead installed to /etc/xdg/AGL, with the defaults to be used named
radio-presets.conf.  A corresponding change to the radio application
to look in that location before trying the appdata directory is also
required.

Bug-AGL: SPEC-2789

Change-Id: Ic5756420e21471096b74b476ad243e44e87ef561
Signed-off-by: Scott Murray <scott.murray@konsulko.com>