Add virtual sllin support and LIN bridging to CAN for demos
To enable seeing the steering wheel LIN messages on the cluster board
for the planned CES demo, a systemd unit that wraps the use of candump
in bridging mode is added when the agl-cluster-support feature is
enabled.  To make sure that the sllin0 interface is always available
for testing without the specific demo LIN hardware, a vcan interface
is set up as sllin0 when the ttyUSB0 device representing the serial
adapter for the LIN transceiver is not present.  Since this virtual
sllin0 interface is very useful for general testing of the steering
wheel event support up through the low-can and signal-composer stack,
it has not been hidden behind agl-cluster-demo-support.

Bug-AGL: SPEC-3049

Change-Id: Idb478e3fe4085859e8704ab0329a341b1a11e0ef
Signed-off-by: Scott Murray <scott.murray@konsulko.com>