Add qt-qrcode recipe for use in settings app
Add a recipe for the qt-qrcode libqrencode wrapper library to allow
using it to display QR codes in the settings app, and update the
settings app dependencies to add it as a dependency. As cleanup,
the qtwebsockets dependency has been dropped from settings since it
no longer requires it.

Bug-AGL: SPEC-2981

Change-Id: Ia0d40ea1151765fa967457bf2ab6624dc7bc259d
Signed-off-by: Scott Murray <scott.murray@konsulko.com>